Reduce number of dispatches needed for GUI text nodes (#12042)
* Do flush after all nodes have been rendered

* Add profiler for dispatches and ignore empty text fields (#12041)

* ignore empty text fields

* ignore dispatches without fields

* Revert "ignore dispatches without fields"

This reverts commit 30f9fe98be9a4c3758d8edd6d9a15623a1414527.

* fix tests

* Add gamesys test

* Review fixes

---------

+++ engine/render/src/render/render_private.h
@@ -162,7 +162,7 @@ namespace dmRender
         uint32_t            m_StringOffset;
         uint32_t            m_OutlineColor;
         uint32_t            m_ShadowColor;
-        uint16_t            m_RenderOrder;
+        uint32_t            m_RenderOrder;
         uint8_t             m_NumRenderConstants;
         bool                m_LineBreak;
         float               m_Width;
